refactor: config edit/update を interface 層に移動
main.rs に直接書かれていた CLI ロジックを正しい層に移動。

- ConfigRepository に save() を追加（domain 層）
- TomlConfigRepository に save() を実装（infrastructure 層）
- config_updater::run() を application 層に新設
- interface/cli/config/edit.rs と update.rs を新設
- main.rs はパス解決とリポジトリのインスタンス化のみ担当
