+++ src/contexts/configuration/application.rs
@@ -1,1 +1,3 @@
 pub mod config_service;
+pub mod config_updater;
+pub use super::domain::repository::ConfigRepository;
