+++ src/contexts/configuration/domain/repository.rs
@@ -2,4 +2,5 @@ use super::config::Config;
 
 pub trait ConfigRepository {
     fn load(&self) -> Config;
+    fn save(&self, config: &Config);
 }
