+++ src/contexts/configuration/infrastructure/toml_loader.rs
@@ -14,6 +14,18 @@ impl ConfigRepository for TomlConfigRepository {
         };
         toml::from_str(&content).unwrap_or_default()
     }
+
+    fn save(&self, config: &Config) {
+        let Some(path) = config_path() else {
+            return;
+        };
+        if let Some(parent) = path.parent() {
+            let _ = std::fs::create_dir_all(parent);
+        }
+        if let Ok(content) = toml::to_string_pretty(config) {
+            let _ = std::fs::write(path, content);
+        }
+    }
 }
 
 // XDG_CONFIG_HOME が設定されていればそちらを優先し、なければ $HOME/.config にフォールバックする。
@@ -23,22 +35,4 @@ pub(crate) fn config_path() -> Option<PathBuf> {
     }
     let home = std::env::var_os("HOME")?;
     Some(PathBuf::from(home).join(".config").join("merge-ready.toml"))
-}
-
-// バージョンが古い場合に設定ファイルをマイグレーションする。最新バージョンならファイルを変更しない。
-pub(crate) fn migrate_config_if_needed(path: &std::path::Path) {
-    use crate::contexts::configuration::domain::config::{CURRENT_VERSION, Config};
-
-    let content = std::fs::read_to_string(path).unwrap_or_default();
-    let mut config: Config = toml::from_str(&content).unwrap_or_default();
-
-    if config.version == CURRENT_VERSION {
-        return;
-    }
-
-    config.version = CURRENT_VERSION;
-    config.fill_defaults();
-    if let Ok(new_content) = toml::to_string_pretty(&config) {
-        let _ = std::fs::write(path, new_content);
-    }
 }
