+++ src/contexts/configuration/interface.rs
@@ -1,0 +1,1 @@
+pub mod cli;
