+++ src/contexts/configuration/interface/cli/config/update.rs
@@ -1,0 +1,5 @@
+use crate::contexts::configuration::application::{self, ConfigRepository};
+
+pub fn run(repo: &impl ConfigRepository) {
+    application::config_updater::run(repo);
+}
